+++ WhirlyGlobeSrc/WhirlyGlobeLib/src/WhirlyGeometry.mm
@@ -113,7 +113,7 @@ bool IntersectLines(const Point2f &p1,const Point2f &p2,const Point2f &p3,const 
 // Clipping planes
 typedef enum {Left,Right,Bottom,Top,Near,Far} ClipPlane;
     
-Eigen::Vector4f intersectPlane(const Vector4f &p1,const Vector4f &p2,ClipPlane plane)
+Eigen::Vector4d intersectPlane(const Vector4d &p1,const Vector4d &p2,ClipPlane plane)
 {
     float t=0.0;
     switch (plane)
@@ -126,12 +126,12 @@ Eigen::Vector4f intersectPlane(const Vector4f &p1,const Vector4f &p2,ClipPlane p
         case Far    : t=(p1.w()-p1.z())/(p2.z()-p1.z()-p2.w()+p1.w());  break;
     }
     
-    Vector4f pt = p1 + (p2-p1)*t;
+    Vector4d pt = p1 + (p2-p1)*t;
     
     return pt;
 }
     
-bool insidePlane(const Vector4f &pt,ClipPlane plane)
+bool insidePlane(const Vector4d &pt,ClipPlane plane)
 {
     switch (plane)
     {
@@ -147,19 +147,19 @@ bool insidePlane(const Vector4f &pt,ClipPlane plane)
     return false;
 }
 
-void ClipHomogeneousPolyToPlane(const std::vector<Eigen::Vector4f> &pts,ClipPlane plane,std::vector<Eigen::Vector4f> &outPts)
+void ClipHomogeneousPolyToPlane(const std::vector<Eigen::Vector4d> &pts,ClipPlane plane,std::vector<Eigen::Vector4d> &outPts)
 {
     outPts.clear();
     for (unsigned int ii=0;ii<pts.size();ii++)
     {
-        const Vector4f &p0 = pts[ii];
-        const Vector4f &p1 = pts[(ii+1)%pts.size()];
+        const Vector4d &p0 = pts[ii];
+        const Vector4d &p1 = pts[(ii+1)%pts.size()];
         bool p0_in = insidePlane(p0,plane);
         bool p1_in = insidePlane(p1,plane);
         // Edge crosses plane
         if (p0_in != p1_in)
         {
-            Vector4f newP = intersectPlane(p0,p1,plane);
+            Vector4d newP = intersectPlane(p0,p1,plane);
             outPts.push_back(newP);
         }
         // 2nd vertex inside, add it
@@ -168,11 +168,11 @@ void ClipHomogeneousPolyToPlane(const std::vector<Eigen::Vector4f> &pts,ClipPlan
     }
 }
     
-void ClipHomogeneousPolygon(const std::vector<Eigen::Vector4f> &inPts,std::vector<Eigen::Vector4f> &outPts)
+void ClipHomogeneousPolygon(const std::vector<Eigen::Vector4d> &inPts,std::vector<Eigen::Vector4d> &outPts)
 {
     if (inPts.size() < 3)
         return;
-    std::vector<Vector4f> pts = inPts;
+    std::vector<Vector4d> pts = inPts;
  
     ClipHomogeneousPolyToPlane(pts, Left, outPts);  pts = outPts;
     ClipHomogeneousPolyToPlane(pts, Right, outPts);  pts = outPts;
@@ -189,21 +189,21 @@ bool RectSolidRayIntersect(const Ray3f &ray,const Point3f *pts,float &dist2)
 }
     
 // Inspired by: http://geomalgorithms.com/a01-_area.html
-float PolygonArea(const std::vector<Point3f> &poly,const Point3f &norm)
+float PolygonArea(const std::vector<Point3d> &poly,const Point3d &norm)
 {
     if (poly.size() < 3)
         return 0.0;
     float area = 0.0;
 
     // Decide which coordinate to ignore
-    Point3f a(std::abs(norm.x()),std::abs(norm.y()),std::abs(norm.z()));
+    Point3d a(std::abs(norm.x()),std::abs(norm.y()),std::abs(norm.z()));
     int coord = (a.x() > a.y()) ? (a.x() > a.z() ? 1 : 3) : (a.y() > a.z() ? 2 : 3);
     
     // Area of the 3D version
     for (unsigned int ii=0;ii<poly.size();ii++)
     {
-        const Point3f &p1 = poly[ii];
-        const Point3f &p2 = poly[(ii+1)%poly.size()];
+        const Point3d &p1 = poly[ii];
+        const Point3d &p2 = poly[(ii+1)%poly.size()];
         switch (coord)
         {
             case 1:
